Fix S3 Version Type
The type of the S3 version depends on whether the resource source is
configured as a versioned file (i.e. the underlying object is versioned,
in which case `version_id` is used) or whether with a regex (i.e. the
underlying object is not necessarily versioned, and the version is
extracted from a regular expression, in which case `path` is used). This
commit fixes this discrepancy so that `path` versions can be pinned.